move comparison to submodule of data
