+++ dcpy/models/data/comparison.py
@@ -1,0 +1,43 @@
+import pandas as pd
+from pydantic import BaseModel, Field
+from typing import TypeVar, Generic
+
+from dcpy.models.base import ModelWithDataFrame
+
+T = TypeVar("T")
+
+
+class Simple(BaseModel, Generic[T]):
+    left: T
+    right: T
+
+
+class Columns(BaseModel):
+    both: set[str]
+    left_only: set[str]
+    right_only: set[str]
+    type_differences: dict[str, Simple[str]]
+
+
+class KeyedTable(ModelWithDataFrame):
+    key_columns: list[str]
+    left_only: set = Field(serialization_alias="Keys found in left only")
+    right_only: set = Field(serialization_alias="Keys found in right only")
+    columns_with_diffs: set[str] = Field(
+        serialization_alias="Columns with changed values for specific keys"
+    )
+    differences_by_column: dict[str, pd.DataFrame] = Field(
+        serialization_alias="Changed values by column"
+    )
+
+
+class SimpleTable(ModelWithDataFrame):
+    compared_columns: set[str]
+    left_only: pd.DataFrame | None
+    right_only: pd.DataFrame | None
+
+
+class Report(BaseModel):
+    row_count: Simple[int]
+    column_comparison: Columns
+    data_comparison: KeyedTable | SimpleTable
